Refactor CandidateAnnotationItem and PinnedAnnotations components to use Source type; update imports and annotations handling for consistency. Add new fetchInstruments function.
