+++ src/scanning/scanningSession/components/ScanningCard/ScanningCardSkeleton.jsx
@@ -2,7 +2,7 @@ import "./ScanningCard.scss";
 import { IonSkeletonText } from "@ionic/react";
 import { THUMBNAIL_TYPES } from "../../../../sources/sources.lib.js";
 import { ThumbnailSkeleton } from "../../../../sources/components/Thumbnail/ThumbnailSkeleton.jsx";
-import { PinnedAnnotationsSkeleton } from "../PinnedAnnotationsSkeleton/PinnedAnnotationsSkeleton.jsx";
+import { PinnedAnnotationsSkeleton } from "../../../../sources/components/PinnedAnnotations/PinnedAnnotationsSkeleton.jsx";
 
 /**
  * @param {Object} props
