+++ src/sources/components/PinnedAnnotations/PinnedAnnotations.scss
