+++ src/sources/components/PinnedAnnotations/PinnedAnnotationsSkeleton.jsx
@@ -1,4 +1,3 @@
-import "../PinnedAnnotations/PinnedAnnotations.scss";
 import { IonSkeletonText } from "@ionic/react";
 
 /**
