+++ src/sources/sources.lib.js
@@ -20,6 +20,7 @@
  * @property {Group[]} groups - Groups the source belongs to
  * @property {Classification[]} classifications - Classifications of the source
  * @property {FollowupRequest[]} followup_requests - Follow-up requests of the source
+ * @property {Annotation[]} annotations - Annotations on the source
  */
 
 /**
@@ -95,6 +96,20 @@
  * @property {number} probability - Probability of the classification
  * @property {string} classification - Classification
  */
+
+/**
+ * @typedef {Object} Annotation
+ * @property {number} id - Annotation ID
+ * @property {string} origin - Annotation origin
+ * @property {string} obj_id - Object ID
+ * @property {{[key: string]: string|number|Array<any>|undefined}} data - Annotation data
+ * @property {number} author_id - Author ID
+ * @property {Group[]} groups - Groups the annotation belongs to
+ */
+
+import { useIonToast } from "@ionic/react";
+import { useCallback } from "react";
+import { Clipboard } from "@capacitor/clipboard";
 
 /**
  * @type {Object<ThumbnailType, ThumbnailType>}
@@ -186,3 +201,72 @@ export function getThumbnailImageUrl(instanceUrl, source, type) {
   }
   return res;
 }
+
+/**
+ * @param {string} group
+ * @param {string} annotationKey
+ * @returns {`${string}/${string}`}
+ */
+export const getAnnotationId = (group, annotationKey) =>
+  `${group}/${annotationKey}`;
+
+/**
+ * @param {string} annotationId
+ * @returns {{key: string, origin: string}}
+ */
+export const extractAnnotationOriginAndKey = (annotationId) => {
+  const lastIndexOfSlash = annotationId.lastIndexOf("/");
+  const origin = annotationId.slice(0, lastIndexOfSlash);
+  const key = annotationId.slice(lastIndexOfSlash + 1);
+  return { origin, key };
+};
+
+
+/**
+ * @param {string|number|undefined} value
+ * @param {number} length
+ * @returns {string|number|undefined}
+ */
+export const concat = (value, length) => {
+  if (typeof value === "string" && value.length > length) {
+    value = value.slice(0, length) + ".."
+  }
+  return value;
+}
+
+/**
+ * @param {string|number|Array<any>|undefined} data
+ * @param {boolean} withIndentation
+ * @returns {string|number|undefined}
+ */
+export const sanitizeAnnotationData = (data, withIndentation) => {
+  if (Array.isArray(data)) {
+    data = JSON.stringify(data, null, withIndentation ? 2 : 0);
+  }else if (typeof data === "boolean") {
+    data = data ? "true" : "false";
+  }
+  return data;
+}
+
+export const useCopyAnnotationLineOnClick = () => {
+  const [present] = useIonToast();
+  return useCallback(
+    /**
+     * @param {string} key
+     * @param {string|number|undefined} value
+     */
+    async (key, value) => {
+      if (value === undefined) {
+        return;
+      }
+      await Clipboard.write({
+        string: `${key}: ${value}`,
+      });
+      await present({
+        message: "Annotation copied to clipboard!",
+        duration: 2000,
+      });
+    },
+    [present],
+  );
+};
